feat(commands): add ping command to display bot latency
This commit adds a new command called "ping" to the commands directory. The ping command is used to display the latency of the bot. It uses the `ctx.client.gateway.latency` property to get the current latency and sends it as a message using an embed. It also calculates the time it took for the message to be sent and adds it to the embed.